# src/go/web/public/docs/openapi.yml
openapi: 3.0.0
info:
  description: phenix API
  version: 0.0.0
  title: phenix
  termsOfService: http://swagger.io/terms/
  contact:
    url: https://github.com/sandialabs/sceptre-phenix
  license:
    name: GNU General Public License v3.0
...
paths:
  "/configs":
    get:
      tags:
        - Configs
      summary: Get all phenix configs
      description: ""
      operationId: getConfigs
      parameters:
        - name: kind
          in: query
          description: limit configs to specified kind (defaults to all)
          required: false
          schema:
            type: string
            enum:
              - all
              - topology
              - scenario
              - experiment
              - image
              - user
              - role
            default: all
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Configs"
    post:
      tags:
        - Configs
      summary: Create new phenix config
      description: ""
      operationId: postConfigs
      requestBody:
        description: phenix config creation parameters
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Config"
          application/x-yaml:
            schema:
              $ref: "#/components/schemas/Config"
          multipart/form-data:
            schema:
              type: object
              properties:
                fileupload:
                  type: string
                  format: binary
      responses:
        "201":
          description: successful config creation
          headers:
            Location:
              schema:
                description: location of newly created config
                type: string
                format: uri
...
servers:
  - url: http://localhost:3000/api/v1
components:
  schemas:
...
    Configs:
      type: object
      properties:
        configs:
          type: array
          items:
            $ref: "#/components/schemas/Config"
    Config:
      type: object
      properties:
        apiVersion:
          type: string
        kind:
          type: string
        metadata:
          type: object
          properties:
            name:
              type: string
            annotations:
              type: object
              additionalProperties:
                type: string
          required:
            - name
        spec:
          type: object
          additionalProperties: true
      required:
        - apiVersion
        - kind
        - metadata
        - spec
...